requirements.txt
flask
yfinance
numpy
pandas
matplotlib.pyplot
scipy